requirements.txt
fastapi==0.115.7
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
PyJWT==2.10.1
SQLAlchemy==2.0.37